pyproject.toml
[tool.poetry]
name = "simple-backend"
version = "0.1.0"
description = ""
authors = ["ylm <1044190862@qq.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.0"
python-multipart = "^0.0.12"
uvicorn = "^0.31.0"
aiomysql = "^0.2.0"
sqlalchemy = "^2.0.35"
pymysql = "^1.1.1"
cryptography = "^43.0.1"
alembic = "^1.13.3"
requests = "^2.32.3"
aiohttp = "^3.10.10"
passlib = "^1.7.4"
bcrypt = "3.2.2"
argon2-cffi = "^23.1.0"
python-dotenv = "^1.0.1"
greenlet = "^3.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
